refactor Renaming, moving code to graveyard.
