--- mule/src/offline/training00_dataset_class.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-
