+++ mule/src/offline/training01_process_datasets.py
@@ -2,10 +2,11 @@
 Manage and analyze the Data Set directory.
 Iterate over each Data Set and check which data elements have been created. 
 
+Standalone script. 
+
 """
 
 import sys, glob, os
-import math
 import json
 import pandas as pd
 #import tensorflow as tf
@@ -491,15 +492,13 @@ class AIDataSet():
         logging.debug("Deleted all .jpg files".format())
 
 #%% Instantiate and load
-#%matplotlib inline
-#plt.ion()
+
 LOCAL_PROJECT_PATH = glob.glob(os.path.expanduser('~/MULE DATA'))[0]
 assert os.path.exists(LOCAL_PROJECT_PATH)
 data1 = AIDataSet(LOCAL_PROJECT_PATH,"20180829 194519")
 data1.augment_df_datetime()
 data1.process_time_steps()
 data1.write_jpgs(overwrite=False)
-
 
 #r = data1.df.head()
 #data1.df.loc[0]
@@ -517,143 +516,6 @@ data1.histogram_throttle()
 data1.plot_turning_frames()
 data1.write_frames()
 
-
-
-#%% DATAGEN
-from tensorflow.python import keras as ks
-#%%
-class MuleDataGenerator(ks.utils.Sequence):
-    """Generates data for Keras"""
-    def __init__(self, indices, dataset, 
-                 batch_size=32, dim=None, n_channels=None, n_classes=15, shuffle=True):
-        """Keras data generator
-        
-        Aggregates the AIDataSet class
-        
-        Attributes:
-            indices (str): The allowed timestamps for data generation
-            dataset (AIDataSet): The dataset object with it's df and npz
-            batch_size : 
-            dim : 
-            n_channels : 
-            n_classes :
-            shuffle :
-        """
-        self.indices = indices
-        self.dataset = dataset
-        self.batch_size = batch_size
-        self.dim = dim
-        self.n_channels = n_channels
-        self.n_classes = n_classes
-        self.shuffle = shuffle
-        self.on_epoch_end()
-        #self.path_frames = path_frames
-        #assert os.path.exists(self.path_frames)
-        #self.path_records = path_records
-        #assert os.path.exists(self.path_records)
-        logging.debug("** Initialize datagen **".format())
-        logging.debug("{} of {} total records used for generation".format(len(self.indices), len(self.dataset.df)))
-        logging.debug("Frames NPZ located at: {}".format(self.dataset.path_frames_npz))
-        logging.debug("{} samples over batch size {} yields {} batches".format(len(self.indices),
-                                                                                   self.batch_size,
-                                                                                   math.ceil(len(self.indices)/self.batch_size),))
-        
-    def __len__(self):
-        """Keras generator method - Denotes the number of batches per epoch
-        """        
-        return int(np.floor(len(self.indices) / self.batch_size))
-    
-    # GET A BATCH!
-    def __getitem__(self, index): 
-        """Keras generator method - Generate one batch of data
-        """         
-        logging.debug("Generating batch {}".format(index))
-        
-        # Generate indexes of the batch
-        batch_indices = self.indices[index*self.batch_size:(index+1)*self.batch_size]
-        #print(batch_indices)
-        
-        # Find list of IDs
-        #list_IDs_temp = [self.list_IDs[k] for k in indexes]
-
-        # Generate data by selecting these IDs
-        X, y = self.__data_generation(batch_indices)
-
-        return X, y
-
-    def on_epoch_end(self):
-        """Keras generator method - Shuffles indices after each epoch
-        """
-        #self.indexes = np.arange(len(self.indices))
-        if self.shuffle == True:
-            # Shuffle is in-place! 
-            np.random.shuffle(self.indices)
-            
-    def __get_npy_arrays(self,batch_indices):
-        """Custom method - get the X input arrays
-        
-        Open the npz file and load n frames into memory
-        """
-        # This is a pointer to the file
-        npz_file=np.load(self.dataset.path_frames_npz)
-        #for k in list_ID_temp:
-        #    npy_records.append(npz_file[k])
-        #X_train = np.array(npy_records)
-        
-        frames_array = np.stack([npz_file[idx] for idx in batch_indices], axis=0)
-        logging.debug("Generating {} frames: {}".format(frames_array.shape[0], frames_array.shape))
-        
-        return frames_array
-    
-    def __get_records(self,batch_indices):
-        """Custom method - get the y labels
-        """
-        # Load the saved records
-        #df_records = pd.read_pickle(self.path_records)
-        # Set the index to match
-        #df_records.index = df_records['timestamp']
-        # Subset
-        #this_batch_steering = df_records.loc[list_IDs_temp]
-        
-        this_batch_df = self.dataset.df.loc[batch_indices]
-        
-        steering_values = this_batch_df['steering_signal'].values
-        
-        #print(steering_values)
-        steering_records_array = self.dataset.bin_Y(steering_values)
-        
-        #df_categorical_steering = df_records['steering_signal']
-        #
-        
-        
-        #records_array = df_records[]
-        logging.debug("Generating {} records {}:".format(steering_records_array.shape[0],steering_records_array.shape))
-        return steering_records_array
-        
-        #raise
-    
-    def __data_generation(self, batch_indices):
-        """Keras generator method - Generates data containing batch_size samples
-        """
-        # X : (n_samples, *dim, n_channels)
-        # Initialization
-        #X = np.empty((self.batch_size, *self.dim, self.n_channels))
-        #y = np.empty((self.batch_size), dtype=int)
-        
-        X = self.__get_npy_arrays(batch_indices)
-        y = self.__get_records(batch_indices)
-        
-        # Generate data
-        #for i, ID in enumerate(list_IDs_temp):
-            # Store sample
-            #X[i,] = np.load('data/' + ID + '.npy')
-
-            # Store class
-            #y[i] = self.labels[ID]
-        #    pass
-
-        return X, y
-    
 
 #%% PROCESS ALL DATASETS!
