+++ mule/src/offline/training04_train_model.py
@@ -1,35 +1,39 @@
-#%% Partition the data
 
-msk = np.random.rand(len(df_records)) < 0.8
+  
+#%% Generate partitions
+#datagen = MuleDataGenerator(data1)
+
+msk = np.random.rand(len(data1.df)) < 0.8
 partition = dict()
-partition['train'] = df_records.index[msk].values
-partition['validation'] = df_records.index[~msk].values
-#labels = range(15)
+partition['train'] = data1.df.index[msk].values
+partition['validation'] = data1.df.index[~msk].values
 
-#%% Generators
-
+#%%
 generator_params = {'dim': (160,120),
           'batch_size': 32,
           'n_classes': 15,
           'n_channels': 3,
           'shuffle': True,
-          'path_frames':os.path.join(LOCAL_PROJECT_PATH,THIS_DATASET,'camera_numpy.zip'),
-          'path_records':os.path.join(LOCAL_PROJECT_PATH,THIS_DATASET,'df_record.pck'),
+          #'path_frames':os.path.join(LOCAL_PROJECT_PATH,THIS_DATASET,'camera_numpy.zip'),
+          #'path_records':os.path.join(LOCAL_PROJECT_PATH,THIS_DATASET,'df_record.pck'),
          }
 
-training_generator = MuleDataGenerator(partition['train'], **generator_params)
-validation_generator = MuleDataGenerator(partition['validation'], **generator_params)
+training_generator = MuleDataGenerator(partition['train'], data1, **generator_params)
+validation_generator = MuleDataGenerator(partition['validation'], data1, **generator_params)
+
 logging.debug("**")
-logging.debug("Data Generators: {} samples over batch size {} yields ~{} batches: {} / {} train/val ".format(len(df_records),
-                                                                                   generator_params['batch_size'],
-                                                                                   math.ceil(len(df_records)/generator_params['batch_size']),
-                                                                                   len(training_generator),
-                                                                                  len(validation_generator)))
+#logging.debug("Data Generators: {} samples over batch size {} yields ~{} batches: {} / {} train/val ".format(len(df_records),
+#                                                                                   generator_params['batch_size'],
+#                                                                                   math.ceil(len(df_records)/generator_params['batch_size']),
+#                                                                                   len(training_generator),
+#                                                                                  len(validation_generator)))
 
-#%% Test the generator class
-
+#%% Testing
 first_batch = validation_generator[0]
 last_batch = validation_generator[len(validation_generator)]
+first_batch = training_generator[0]
+last_batch = training_generator[len(validation_generator)]
+
 
 #%% The model
 
@@ -80,9 +84,36 @@ sum(df_records['steering_pred_argmax'] == df_records['steering_signal_argmax'])/
 df_records['steering_pred_signal'] = df_records['steering_pred_signal'].apply(lambda x: x*-1)
 logging.debug("Steering signal inverterted - WHY?".format())
 
-#%% =============================================================================
-# Plot analysis of model
-# =============================================================================
+#%% Plot analysis of model
+
+def plot_hist(history_dict, accuracy_name, model_title):
+    #fig = plt.figure(figsize=(5,4))
+    #fig=plt.figure(figsize=(20, 10),facecolor='white')
+
+    #f, (ax1, ax2) = plt.subplots(1, 2, figsize=(20, 5),sharey=False,facecolor='white')
+    bgcolor = '0.15'
+    bgcolor = 'white'
+    f, (ax1, ax2) = plt.subplots(1, 2, figsize=(20, 5),sharey=False,facecolor=bgcolor)
+    
+    ax1.plot(history_dict['epoch'],  history_dict['history']['loss'],label="Train")
+    ax1.plot(history_dict['epoch'],  history_dict['history']['val_loss'],label="CV")
+    ax1.set_title("Loss function development - Training set vs CV set")
+    ax1.legend(loc='upper right')
+    ax1.set_xlabel('Epochs')
+    ax1.set_ylabel('Values')
+    
+    ax2.plot(history_dict['epoch'],  history_dict['history'][accuracy_name],label="Train")
+    ax2.plot(history_dict['epoch'],  history_dict['history']['val_'+accuracy_name],label="CV")
+    ax2.set_title("Accuracy development - Training set vs CV set")
+    ax2.legend(loc='upper right')
+    ax2.set_xlabel('Epochs')
+    ax2.set_ylabel('Values')
+    
+    plt.suptitle(model_title, fontsize=16)
+    
+    plt.show()
+
+
 # Plot training history
 #%matplotlib inline
 #model_title = "10 Epochs"
@@ -109,9 +140,7 @@ these_records = get_full_records(frames_npz, df_records, these_indices)
 plot_frames(these_records)
 
 
-#%% =============================================================================
-# Saving the model
-# =============================================================================
+#%% Saving the model
 # Save the model weights, architecture, configuration, and state
 # serialize weights to HDF5
 path_model_h5 = os.path.join(LOCAL_PROJECT_PATH,THIS_DATASET,'model',this_timestamp + ' model.h5')
